Adds ability to force subscription bill dates to a certain day of the month.
Use Case: Companies that produce a physical subscription product on a
monthly basis (example: print magazine) may prefer to collect
subscription funds on the same day each month in order to both have
funds available to purchase materials and to determine the quantity of
product to produce that month.

For Example: Print Magazine company ships March's magazine during last
week of February. Magazine content must be finalized and sent to printer
by February 15th. Therefore, the print magazine company needs to charge
all their subscribers on the 10th in order to be able to both clear
funds and send to printer by the 15th and to know exactly how many
magazines to print.