Fix for adding multiple customizations to a single SKU in the cart
Fixes this issue posted to main Slatwall repo.

1. Create a Product with an Order Item Attribute Set for "Custom Name"

2. Add a product to the cart with the Custom Name "Dan"

3. Add a second instance of the same product with the Custom Name "Greg"

Expected Result:

The order will contain two line items, one with the AttributeValues "Dan", and the other with the AttributeValues "Greg"

Actual Result:

One SKU is added to the cart with the AttributeValue "Dan", and a quantity of 2.

Greg stated in the forums that the issue is in the OrderService.cfc lines 360-365. I'm working on a fix, but haven't come to the solution yet.