Update HibachiReport.cfc
Proposing update to conditional logic validating and compiling report data for charts. 

We were seeing the charts almost never actually plotted any data points. Looking into this, I think I found some issues in the logic used to compile the array that gets returned. I tweaked the conditionals as well as added an isDefined for the field that was causing errors. 

These might be Railo-specific so more testing on ACF should be done, however, this fix has our charts finally working as expected.