Added slatwall service method for retrieving resized image based profilename and skuids.
Added logic to use that deferred promise in the swresizedimage
directive.

Made a small change to orderfulfillment-shippinglabel to use a ! in the
ng-if statement (so N/A only displays when the address is not
available).

Renamed swoiimage swresizedimage (needs to still be made more generic).

Refactored the image directive to use a templateUrl rather than a
template string.