Update AttributeValue.cfc
We've been working with a webservice that returns a binary file, that we pre-upload from the response. The file is needed to be viewable as a custom Attribute. The condition to check for a 'form' key has recently broken this function. We propose a mod to check for a pre-existing file path so that a File Attribute can be used in both modes.