Fixes issue where order was being defined as struct
when the
```
var = order ...
```
logic wasn't firing Lucee and Railo were still picking up a value for order, but it was a struct of order.orderID, I think the data might have been getting passed into the arguments scope.  Not sure if the error exists on ACF, but this should be fine either way because order is always passed around by reference so no-need to set it into an additional variable.