Fixed the edit function
The .toggle() function has been deprecated and no worked correctly. I
replaced it with a simple function to swap element visibility. I placed
the function in the footer include. Is there maybe a better place for
it?

There is still an issue with the delete account address and payment
method functions. The page reloads, but instead of removing the address
it loads with edit form visible. Hitting cancel or save then shows the
display address.