Update Shipping Service
This adds some checks to make sure that the shipping method options that are being added are allowed to be added. Previously, it was adding both manual and worldwide shipping methods to the fulfillment that it should not have been. Also, I updated the condition that decides when a new shipping method rate should be selected by default. Before, it would auto select if one and only one was available. I changed it to auto select the first one if at least one is available.