Gift Card Changes: Work in progress
Added ability for gift cards with manual codes to be added when orderItem is added to order.
Updated auto-fulfillment logic to be able to handle the different types of gift card creation processes.
Added a lot of missing gift card validation.
Moved existing GiftCard global settings to offer more granular control at sku/product/productType level.
Some refactoring and decoupling to allow for standardized validation to be properly handled

* Lots of updates to logic in processes:
 - OrderService.Order_AddOrderItem
 - OrderService.Order_AddOrderItemGiftRecipient
 - OrderService.Order_PlaceOrder
 - OrderService.OrderDelivery_Create
 - OrderService.OrderDeliveryItem_CreateGiftCards (new)

* UI Changes to:
 - preprocessorder_addorderitem.cfm
 - preprocessorderdelivery_create.cfm
 - skusettings.cfm (productType, product, & sku entities)

* Lots of gift card related validation that was missing, some modifications needed to be move logic into the standardized process methods for validation.
We are now checking:
 - to make sure codes were present in the data under the right scenarios
 - for no duplicate codes with data being processed
 - the database to make sure gift card is not already in-use
 - to make sure correct gift card code character length is valid