Update global.js
This fixes an issue where selecting a card was not opening any forms to continue.
I tested that this opens the correct form. That it only opens the correct forms and not al of them.
I tested that if you change the selector back to blank, the forms disappear. I tested on other modals as well.