Update order amounts with taxes
Fixes an issue where taxes don't calculate for return items on a sales order because the order was paid for previous to the return item being added. The current check only recalculates if the payment due is great than 0, but in the case where you are adding a return sku on  paid order, the payment amount due will be -negative. Updating the check to recalculate whenever the payment is not zero, fixes because if the custom is owed or the customer owes, taxes are applied. But when its balanced (0), it does not recalculate.